Add durable attribute to http api. when durable commits are SYNC, otherwise NOSYNC. add tests

+++ core/src/main/java/io/questdb/cutlass/http/processors/TextImportProcessor.java
@@ -144,7 +144,7 @@ public class TextImportProcessor implements HttpRequestProcessor, HttpMultipartC
             CharSequence commitLagChars = rh.getUrlParam("commitLag");
             if (commitLagChars != null) {
                 try {
-                    long commitLag = Numbers.parseInt(commitLagChars);
+                    long commitLag = Numbers.parseLong(commitLagChars);
                     if (commitLag >= 0) {
                         transientState.textLoader.setCommitLag(commitLag);
                     }
